+++ docker/apl/run.sh
@@ -1,2 +1,0 @@
-printf %s "$1" > program.apl
-apl --OFF -s -f program.apl || true