+++ docker/elixir/run.sh
@@ -1,2 +1,0 @@
-printf %s "$1" > program.exs
-elixir program.exs || true