+++ docker/fsharp/run.sh
@@ -1,2 +1,0 @@
-printf %s "$1" > program.fs
-fsharpc --optimize- program.fs >/dev/null && mono program.exe || true