+++ docker/go/run.sh
@@ -1,3 +1,0 @@
-export GOCACHE=/tmp/"$CODEDIR"/cache
-printf %s "$1" > program.go
-go run program.go || true