+++ docker/lua/run.sh
@@ -1,2 +1,0 @@
-printf %s "$1" > program.lua
-lua5.3 program.lua || true