+++ docker/ocaml/run.sh
@@ -1,2 +1,0 @@
-printf %s "$1" > program.ml
-ocamlopt -cclib --static -o program program.ml && ./program || true