+++ docker/perl/run.sh
@@ -1,2 +1,0 @@
-printf %s "$1" > program.pl
-perl program.pl || true