+++ docker/php/run.sh
@@ -1,2 +1,0 @@
-printf %s "$1" > program.php
-php program.php || true