+++ docker/racket/run.sh
@@ -1,2 +1,0 @@
-printf %s "$1" > program.rkt
-racket program.rkt || true